poc/backend/requirements.txt
fastapi
uvicorn
sqlmodel
passlib[bcrypt]
python-jose
boto3
qdrant-client
requests
